UUID: int2long clean up

+++ app/src/main/java/com/money/manager/ex/budget/BudgetsActivity.java
@@ -67,7 +67,7 @@ public class BudgetsActivity
         // Handle action bar item clicks here. The action bar will
         // automatically e clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
-        long id = item.getItemId();
+        int id = item.getItemId();
 
         //noinspection SimplifiableIfStatement
         if (id == R.id.action_settings) {
